task #15: Update code to use the recording tree where possible
Update the recorder and the session results object to use the recording
tree instead of the flat list to avoid loosing the tree information
unnecessarily. To facilitate testing, a new property was added to the
recorders to get back the flat list of results instead of the tree. The
tests were also updated to generate mock tree data instead of mock list
data.