task #15: Fix bug in tree search
Fix sneaky bug in the recording tree Find method that causes the
incorrect node to be returned. The fix is the return the actual
result instead of child that produced a non-null result.
Update tests to ensure the object returned by the Find method
matches the data being searched for, not just that the result is
non-null.