task #15: Update the nested result formatter to also pad values
Update the nested result formatter to print the values similar to the
padded result formatter in that the values are lined up in columns
by adding spaces to make sure everything lines up.
This required flattening out the result tree to find the longest value
and longest count to pad the string correctly.
Update the test to the new format.