task #19: Fix nesting in performance recorder
Update how the performance recorder nests results by putting the actual
tree node on the method stack instead of only the method definition. The
reason for this is that the definition of the parent method alone is not
enough to ensure the method gets placed correctly in the tree, as the
parent method can be in several places in the tree. By stacking the actual
tree node, when a new method is registered it is guaranteed to go in the
correct spot.
This required refactoring the attribute class to stack the tree nodes and
the performance recorder itself to return the created nodes when a method
is registered for the first time.