task #17: Update target framework of library
Update the target framework from `netcoreapp2.2` to `netstandard2.0` to
provide more compatibility to projects using the library.
This change requires some code updates since the code was using a couple
of functions that no longer exist as part of net standard.
Update the csproj file with package information to allow packaging the
library using visual studio.