task #20: Update code to use new logger
Update the result recorder and the recording attribute to make use of the
new logger if it is available.
Add a new inactive logger to always be used as a default if no injected
logger is available. This ensures the logger is never null, even if the
client code sets the logger to null.