task #23: Add new depth property to recording result
Update the recording result with a new Depth property to track how deep
the method being recorded is in the recording tree.
Update the recording constructor to remove the ability to add the first
result and refactor code using that constructor - this is to later support
passing the depth down into the recording data.