task #27: Add property to set the number of decimal places in output
Add a new property to the session results object to allow configuring
the number of decimal places in the output strings.
Refactor how the session result object handles formatters to create a new
formatter and pass down the properties in the constructor instead of
re-using the same object over and over again and re-setting the properties.