task #27: Update formatter to ensure valid number of decimal places
Update the string formatter to ensure the number of decimal places used
is not less than 0 and not greater than 3. This is done by setting it to
the closest valid value if the given number is outside the range. Doing
this and logging an error instead of throwing an exception because it
can easily be handled and should not cause client code to crash.