task #28: Add method to record action duration
Add a new method on the library API to record the duration of a given
action. This is used by passing in a name and an Action object, the code
will run the provided action and record its duration.
Split out code from the performance logging attribute into a static worker
to allow it to be re-used for this functionality.