#44: Package project as first step in CI
Update the gitlab CI pipeline to build and package the project as the
first step of the build instead of only packaging the library for tags
and on the develop branch. This makes it easier to download test versions
of the library for testing and helps catch all possible issues before
the code is merged.