build: add build and version update make commands
Update the make file to include the build and update version commands.
Remove the call to update the version in the build script. The update
version command now needs to be called before the build command.
Add exec permission to the build scripts.
Update the gitlab CI script to use make commands.