feat: refactor recording session result interface (#32)
Refactor the `RecordingSessionResult` interface to expose the flattened
and combined results as a property instead of a method. Add documentation
to the new property to explain that it combines results for the same method
into a single result.
Add documentation to the `Flatten` and `FlattenAndCombine` methods to help
explain when to use one or the other.

BREAKING CHANGE: Remove `Count` property from `IRecordingSessionResult`, use `FlatRecordingData.Count` instead.
BREAKING CHANGE: Remove `FlatData()` method from `IRecordingSessionResult`, use the `FlatRecordingData` property instead.