ci: build documentation site in ci pipeline (#50)
Update the Gitlab CI/CD pipeline to build out the documentation site as
well as building and testing the library.
The pipeline definition was modified to use a template stage for all the
dotnet stages. This provides the benefit of a shared `before_script`
section without running those commands for the docs stage.